Create Minimum Amount of Time to Collect Garbage.cpp
here is  two string --> garbage = ["G","P","GP","GG"] and 
travel = [2,4,3]
in garbage string G stands for Glass Garbage, P stands for Paper Garbage, also there is a possibility of M which stands for Metal garbage. 

there are 3 trucks for collecting 3 different types of garbage. 
It will take 1 minute to collect every garbage.
and the traveling array stands for the timing to travel from Grabge's one index to another.  So 0 to 1--> 2min, 1 to 2--> 4min, and 2 to 3--> 3min.

So we have to calculate the total minimum number of minutes needed to pick up all the garbage.